Move EndpointsController to its own package
Avoids recursive loop with endpoint registry defined
