+++ pkg/service/endpoints_controller_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package endpoint
+package service
 
 import (
 	"encoding/json"
